Fixes #182. Moved IP classes to net package.
