+++ mango-core/src/main/java/org/calrissian/mango/net/IP.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.calrissian.mango.domain.ip;
+package org.calrissian.mango.net;
 
 
 import java.io.Serializable;
