+++ mango-core/src/main/java/org/calrissian/mango/net/IPv6.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.calrissian.mango.domain.ip;
+package org.calrissian.mango.net;
 
 
 import com.google.common.collect.DiscreteDomain;
 import com.google.common.collect.Range;
-import org.calrissian.mango.net.MoreInetAddresses;
 
 import java.math.BigInteger;
 import java.net.Inet6Address;
