+++ mango-core/src/main/java/org/calrissian/mango/types/LexiTypeEncoders.java
@@ -19,8 +19,8 @@ package org.calrissian.mango.types;
 import com.google.common.primitives.UnsignedInteger;
 import com.google.common.primitives.UnsignedLong;
 import org.calrissian.mango.domain.entity.EntityIdentifier;
-import org.calrissian.mango.domain.ip.IPv4;
-import org.calrissian.mango.domain.ip.IPv6;
+import org.calrissian.mango.net.IPv4;
+import org.calrissian.mango.net.IPv6;
 import org.calrissian.mango.types.encoders.lexi.*;
 
 import java.math.BigDecimal;
