+++ mango-core/src/main/java/org/calrissian/mango/types/encoders/simple/IPv4Encoder.java
@@ -16,7 +16,7 @@
 package org.calrissian.mango.types.encoders.simple;
 
 
-import org.calrissian.mango.domain.ip.IPv4;
+import org.calrissian.mango.net.IPv4;
 import org.calrissian.mango.types.encoders.AbstractIPv4Encoder;
 
 import static com.google.common.base.Preconditions.checkNotNull;
