+++ mango-core/src/main/java/org/calrissian/mango/types/encoders/simple/IPv6Encoder.java
@@ -16,7 +16,7 @@
 package org.calrissian.mango.types.encoders.simple;
 
 
-import org.calrissian.mango.domain.ip.IPv6;
+import org.calrissian.mango.net.IPv6;
 import org.calrissian.mango.types.encoders.AbstractIPv6Encoder;
 
 import static com.google.common.base.Preconditions.checkNotNull;
