+++ mango-core/src/test/java/org/calrissian/mango/types/LexiTypeEncodersTest.java
@@ -18,8 +18,8 @@ package org.calrissian.mango.types;
 import com.google.common.primitives.UnsignedInteger;
 import com.google.common.primitives.UnsignedLong;
 import org.calrissian.mango.domain.entity.EntityIdentifier;
-import org.calrissian.mango.domain.ip.IPv4;
-import org.calrissian.mango.domain.ip.IPv6;
+import org.calrissian.mango.net.IPv4;
+import org.calrissian.mango.net.IPv6;
 import org.junit.Test;
 
 import java.math.BigDecimal;
