+++ mango-json/src/test/java/org/calrissian/mango/json/ser/NodeSerializerTest.java
@@ -20,7 +20,7 @@ import org.calrissian.mango.criteria.builder.QueryBuilder;
 import org.calrissian.mango.criteria.domain.EqualsLeaf;
 import org.calrissian.mango.criteria.domain.Node;
 import org.calrissian.mango.criteria.domain.RangeLeaf;
-import org.calrissian.mango.domain.ip.IPv4;
+import org.calrissian.mango.net.IPv4;
 import org.calrissian.mango.json.MangoModule;
 import org.junit.Test;
